Creation of CumulativeMeta, simplification of CumulativeSum, and moving add kernel to new internal file
